Add async test case switch
Tets-jstransformer applies all tests to each method it can find,
regardless of the method being asynchronous or not. This is a problem
for tests that require a method to be asynchronous. These test cases
should only be applied to the asynchronous methods. This commit checks
for the text '_async' in the folder name. If it can find it, only the
asynchronous test cases are performed. Asynchronous methods are tested
on synchronous test cases, because that way around it should work,
regardless of a method being asynchronous.

`checkFunctionOutput` is now asynchronous. This is done, because it is
not a given that `template.fn(locals)` returns a string. Hence this test
case can fail because `template.fn(locals).trim()` returns an error. By
wrapping it in a `Promise.resolve()` both the situations where
`template.fn(locals)` returns a string or a Promise are covered. This
is also how the jstranformer library works, so the tests now better
mimmick the behavior of the actual library.